refactor: enum imports update
