+++ src/components/ui/layout-components/actions/actions.tsx
@@ -3,7 +3,7 @@ import { ComponentPropsWithoutRef, ReactNode } from 'react'
 import { EditOutline, PlayCircleOutline, TrashOutline } from '@/assets/components/svgIcons'
 import { getActionButtons } from '@/components/ui/layout-components/actions/utils/utils'
 import { Button } from '@/components/ui/primitives'
-import { ACTIONS, VARIANT } from '@/shared/enums/actions'
+import { ACTIONS, VARIANT } from '@/shared/enums'
 import { FlexContainer } from '@/shared/ui/flex-container'
 import clsx from 'clsx'
