+++ src/components/ui/layout-components/deck-table/deck-table.tsx
@@ -15,8 +15,7 @@ import {
   TableHeaderCell,
   TableRow,
 } from '@/components/ui/primitives'
-import { VARIANT } from '@/shared/enums/actions'
-import { RATIO } from '@/shared/enums/ratio'
+import { RATIO, VARIANT } from '@/shared/enums'
 import { FlexContainer } from '@/shared/ui/flex-container'
 import clsx from 'clsx'
