+++ src/components/ui/layout-components/deck-title/deck-title.tsx
@@ -1,6 +1,6 @@
 import { SettingsDropdown } from '@/components/ui/layout-components'
 import { Image, Typography } from '@/components/ui/primitives'
-import { RATIO } from '@/shared/enums/ratio'
+import { RATIO } from '@/shared/enums'
 import { FlexContainer } from '@/shared/ui/flex-container'
 import clsx from 'clsx'
