+++ src/components/ui/layout-components/learn-card/learn-card.tsx
@@ -5,7 +5,7 @@ import dummyQuestionCover from '@/assets/webp/dummy-question-cover.webp'
 import { SelfRateForm } from '@/components/forms/self-rate-form/self-rate-form'
 import { CardListExample } from '@/components/ui/layout-components/deck-table/deck-table.mock'
 import { Button, Card, Image, Typography } from '@/components/ui/primitives'
-import { RATIO } from '@/shared/enums/ratio'
+import { RATIO } from '@/shared/enums'
 import { FlexContainer } from '@/shared/ui/flex-container'
 import clsx from 'clsx'
