+++ src/components/ui/primitives/image/image.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta } from '@storybook/react'
 
 import myImage from '@/assets/webp/react-logo.webp'
 import { Image } from '@/components/ui/primitives/image/image'
-import { RATIO } from '@/shared/enums/ratio'
+import { RATIO } from '@/shared/enums'
 
 const meta = {
   component: Image,
